En teoria RR ya sirve, voy a probar mas
